+++ packages/app-files-web/tailwind.d.ts
@@ -1,0 +1,7 @@
+declare module "tailwindcss/src/util/dataTypes" {
+  export function normalize(value: string): string;
+}
+
+declare module "tailwindcss/src/util/escapeClassName" {
+  export default function (value: string): string;
+}
